Move geofence into geo builder
