+++ synctools/src/main/kotlin/at/bitfire/synctools/mapping/jtx/builder/GeoBuilder.kt
@@ -5,9 +5,11 @@
 package at.bitfire.synctools.mapping.jtx.builder
 
 import android.content.Entity
+import at.bitfire.ical4android.JtxICalObject
 import at.techbee.jtx.JtxContract
 import net.fortuna.ical4j.model.component.CalendarComponent
 import net.fortuna.ical4j.model.property.Geo
+import net.fortuna.ical4j.model.property.XProperty
 import kotlin.jvm.optionals.getOrNull
 
 class GeoBuilder : JtxObjectEntityBuilder {
@@ -15,5 +17,8 @@ class GeoBuilder : JtxObjectEntityBuilder {
         val geo = from.getProperty<Geo>(Geo.GEO).getOrNull()
         to.entityValues.put(JtxContract.JtxICalObject.GEO_LAT, geo?.latitude?.toDouble())
         to.entityValues.put(JtxContract.JtxICalObject.GEO_LONG, geo?.longitude?.toDouble())
+
+        val geofenceRadius = from.getProperty<XProperty>(JtxICalObject.X_PROP_GEOFENCE_RADIUS).getOrNull()
+        to.entityValues.put(JtxContract.JtxICalObject.GEOFENCE_RADIUS, geofenceRadius?.value?.toIntOrNull())
     }
 }
